refactor(cli): extract cli from core module
