+++ mwc/cli/cli.py
@@ -4,9 +4,7 @@ from mwc.core.cfg import load_config
 from mwc.core.db.models import init_db
 from mwc.core.logger import config_logger
 import logging
-from mwc.bots.cli import tweet_movie
-from mwc.core.movies.cli import sync_tmdb
-from mwc.core.subtitles.cli import download_missing_subtitles
+from mwc.cli.commands import tweet_movie, download_missing_subtitles, sync_tmdb
 
 CONFIG = load_config()
 log = logging.getLogger(__name__)
