+++ setup.py
@@ -31,7 +31,7 @@ setuptools.setup(
     python_requires='>=3.6',
     entry_points={
         'console_scripts': [
-            'mwc=mwc.cli:main',
+            'mwc=mwc.cli.cli:main',
         ],
     },
 )
